requirements.txt
fredapi==0.5.2
holidays==0.63
matplotlib==3.10.0
numpy==2.2.1
pandas==2.2.3
plotly==5.24.1
python-dotenv==1.0.1
python_dateutil==2.9.0.post0
scipy==1.14.1
seaborn==0.13.2
statsmodels==0.14.4
streamlit==1.24.0
tensorflow==2.18.0
ydata_profiling==4.12.1
yfinance==0.2.51
